// portfolio/src/sanity/types.ts
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type GET_ARTICLES_QUERYResult = Array<{
  title: string | null;
  description: string | null;
  publishedAt: string | null;
  slug: Slug | null;
}>;